refactor(plugins): plugin-env seam for the settings dialog declarations
Move loadPluginEnvDecls (tongflow.plugin.json reader) behind
@ext/plugin-env following the plugin-registry seam pattern: the default
backend in src/ext-default/plugin-env.ts keeps the filesystem scan;
a cloud shell substitutes a build-time baked declaration list via
src/ext/plugin-env.ts. No behavior change for desktop/self-host.
